+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Master
+## 0.0.8
 
 - Add Atlas to `ProductMenuItems`. ([#48](https://github.com/mapbox/dr-ui/pull/48/))
 - Show scroll bar automatically in `PageLayout` when the sidebar contents overflow vertically. ([#54](https://github.com/mapbox/dr-ui/pull/54))